Style: Detail & Link 컴포넌트 분리
