+++ babel.config.js
@@ -1,3 +1,6 @@
 module.exports = {
-  presets: [["@babel/preset-react"], "@babel/preset-typescript"],
+  presets: [
+    ["@babel/preset-react", { runtime: "automatic" }],
+    "@babel/preset-typescript",
+  ],
 };
